Minor optimizations and warnings fixed

+++ BrickBreak/src/Ball.cpp
@@ -6,7 +6,7 @@ Ball::Ball()
 Ball::Ball(const glm::vec2& pos, float radius, glm::vec2 velocity, Texture* texture)
     :Object(pos, glm::vec2(radius * 2.0f, radius * 2.0f), texture, 0.0f, glm::vec4(1.0f), velocity), radius(radius), stuck(true), initialVelocity(velocity) {}
 
-glm::vec2 Ball::move(double delta, unsigned int windowWidth)
+glm::vec2 Ball::move(double delta, float windowWidth)
 {
     if (!stuck)
     {
